requirements.txt
Pillow==10.1.0
thread6==0.2.0